+++ CHANGELOG.md
@@ -56,3 +56,9 @@ Initial release with the following features:
 - Added topics
 - Added ability to override dio instance in CesiumHttpService
 - Added ActionResource to handle async actions cleanly
+
+## 2.0.0
+
+- Moved all constructors to have named parameters
+- Added manage function in ManagedListenerMixin
+- Added onSuccess and onFailure to ActionResource